requirements.txt
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
dhis2.py==2.3.0
greenlet==1.1.2
idna==3.3
logzero==1.7.0
pycodestyle==2.8.0
Pygments==2.12.0
pyodbc==4.0.32
python-dotenv==0.20.0
requests==2.28.1
toml==0.10.2
urllib3==1.26.9
